+++ packages/atomic/vitest.config.ts
@@ -12,6 +12,7 @@ export default defineConfig({
     'import.meta.env.RESOURCE_URL': `"${resourceUrl}"`,
     __ATOMIC_VERSION__: `"${packageJson.version}"`,
     __HEADLESS_VERSION__: `"${packageJson.dependencies['@coveo/headless']}"`,
+    'process.env': {},
   },
   server: {
     port: port,
@@ -59,6 +60,9 @@ export default defineConfig({
       instances: [
         {
           browser: 'chromium',
+          context: {
+            actionTimeout: 1000,
+          },
         },
       ],
     },